refactor(canvas): extract home-canvas hooks from FreeformCanvasView
Address the React Doctor `no-giant-component` warning by moving the
home-canvas reset flow and the canvas→host navigation switch into focused,
colocated hooks (`useHomeCanvasReset`, `useCanvasNavigation`). FreeformCanvasView
now renders and delegates the logic, dropping ~60 lines from the component.

No behavior change.

Generated-By: PostHog Code
Task-Id: 5f5858a1-f09c-41bb-b336-c2db1266259c
